Extract tests for isArrayBuffer to own file
